Move interface comment onto interface.
In generated code, before, the interface comment was put in the type
declarations. However, the hidl-gen implementation implemented
interfaces differently than other types, and so the
emitTypeDeclarations method when called on a root scope which only
contained an interface would emit the doc comment but not emit
an interface definition (which is emitted by generateCpp instead).
This caused the comment to appear in the middle of the file and
annotate the next item in the file.

The same bug existed in the Java implementation.

Bug: 115508949
Test: manual
Change-Id: Ib60bb047d29293852006cb82d6de412ab82b8c76