hidl_test*: working around breaking build changes.
The build no longer supports sync/run for some types of tests for various reasons. Until we are
setup w/ atest, restoring the build so that our existing testing scripts work.

Notice that hidl_test_java no longer works as it was because of Soong's strong enforcement.
Instead, scripts have been updated with:
    mkdir -p $ANDROID_PRODUCT_OUT/data/framework/hidl_test_java.jar || return 1
    cp $ANDROID_PRODUCT_OUT/testcases/hidl_test_java_java/arm64/hidl_test_java_java.jar \
       $ANDROID_PRODUCT_OUT/data/framework/hidl_test_java_java.jar || return 1

Bug: 127714188
Test: m hidl_test && adb sync data && adb shell /data/nativetest64/hidl_test
Test: system/tools/hidl/test/run_all_device_tests.sh
Change-Id: Ic8004948dc3f87dbbc597b14eb379d9116eefc33