HAL query scripts for unreleased interfaces.
Moving some old scripts that have been laying around here. They figure
out FQNames of HIDL interfaces from the file system.

Specifically, 'aosp-changes-report' can be used to see all of the
interfaces which are unreleased and in the current branch (if they are
used on a device, you should also see 'N' in the lshal 'R' column).

Fixes: 138307418
Test: source $ANDROID_BUILD_TOP/system/tools/hidl/scripts/hal-queries.sh
(and running all added scripts)
Test: hidl-gen -Lcheck $(aosp-interfaces)

Change-Id: Ief6cf5c883d55676a9cb76b135a5b89baf1fa8d7