Modify StringHelper::Tokenize to fix HidlLintTest
This was broken by aosp/1093499. VAL2 is now being tokenized as VA_L2
(which seems wrong). This changes StringHelper::Tokenize such that
numbers are treated as upper case characters.

VAL2 is now tokenized as VAL2

We should add this test into TEST_MAPPING so that we can catch this in
presubmit.

Bug: N/A
Test: ./test/run_all_host_tests.sh
Change-Id: I7c0596da88980e585c1fc3ed7178f356527c3841