[hidl-lint] Allow oneway lint to run on IBase
This doesn't change anything since the code still depends on
userDefinedMethods. I think this is more to do with the (incorrect?)
implementation of `methods()` than the lint code.

Methods always contains the reserved methods even though it shouldn't
since IBase (which is in the superTypeChain) would contain those.
Because of this userDefinedMethods has to be used.

Bug: N/A
Test: ./test/run_all_host_tests.sh
Change-Id: I605b6113a6c7a61db21113182583743ed13e637e