Fix VectorType Java getJavaType.
Previously, this method used 'getJavaType' instead of 'getJavaTypeClass'
if the element is a templated type. However, 'isTemplatedType' refers to
whether a type is templated in the HIDL language, not whether it is
templated in the Java language. So, bitfield<*> which is templated in
HIDL (but a primitive in Java) would emit things like 'ArrayList<int>'.

Since Vector in HIDL Java is not allowed to contain any type which is
templated in the output, simply changing this to always return
getJavaType (vec<int[]>, vec<vec<int>>, fmq_*<*> are all disallowed in
Java).

Fixes: 141187928
Test: build all HIDL Java interfaces
Test: build IBaz which now relies on this type to exist
Change-Id: Ie238becf548f7e00405826aef48e395757e5bdf8