hidl-gen need to support generating java interface for specified hal file
Sometimes user creates a few hal files in his vendor/*/interfaces.
for example.
vendor/synaptics/interfaces/foo/1.0/
         |- IFoo.hal
         |- IBuz.hal
         |- IBuss.hal
And now user wants to generate only IFoo java interface from IFoo.hal, but not IBuz and IBuss interfaces.
So I execute below command to do that:
$hidl-gen -o path/to/java/output -Ljava -randroid.hidl:system/libhidl/transport
    -rcom.synaptics:path/to/vendor/interfaces com.synaptics.foo@1.0::IFoo

But after I check the path/to/java/output, I can see all the interfaces (IFoo, IBuz, IBuss)
of the same package are generated java file in the output folder. Obviously this is not what
user expects.

The root cause is in function OutputHandler::appendTargets() line 191,
if the fqName.isFullQualified() is true, I think it should only append the fqName+types to
the target and then return, don't need to append all the interfaces in the same package.

Change-Id: I8399eaa3320dde73d3592f13c2c9fb8aff98cc5f