Verbose mode: avoid extra enforcements
hidl-gen clears interface hashes when generating interfaces so that VTS
has a chance of detecting an unfrozen interface (this is the only way we
can recommend interface owners freeze their interfaces and are therefore
aware of when they make an interface-breaking change). However, it isn't
intentional for '-v' to always invoke this since the '-Lhash' output
mode requires the hash is always outputted, even if the interface is
unfrozen (so that '-Lhash' can be used to freeze an interface).

Fixes: 145021290
Test: 'hidl-gen -Lhash' on unfrozen interface w/ and w/o '-v' always
  shows a hash

Change-Id: I968b03ec600893bcfa61ed8f7ca13c16fffa893f