hidl2aidl: avoid overwriting ignored method versons
We were previously pushing a pointer to a method in a map into a vector
of ignored methods. We immediately overwrote the values in the map that
the poniter was pointing to.

Now we push a copy of the ignored method into the vector instead of a
pointer. Using the original pointer, we overwrite the values in the map
with the new method.

Bug: 151878846
Test: m SANITIZE_HOST=address hidl2aidl && hidl2aidl -o . android.hardware.camera.device@3.5
Change-Id: Id7ebf69b5a3d51840dd1af9888582baf7153d4eb