Flatten compound types that include older versions of themselves
In HIDL, new versions of structs could add a field of the older version
to keep all of the functionailty.
When converting these structs, we will now expand the older version
inside the newest struct so it will contain all of the fields.

nfc@1.2 had a good example of this in NfcConfig.hal.

Test: m hidl2aidl && hidl2aidl -o . android.hardware.nfc@1.2
Test: cd system/tools/hidl/hidl2aidl && mm
Bug: 154053350
Change-Id: I2abc2d9b9d94ab93e8080bd34492bf2b348f572e