hidl2aidl: support only one interface
In preparation for emitting an Android.bp file at the base of the output
file structure. hidl2aidl is expected to run separately for separate
HIDL packages anyway because they belong in different places.

For instance, this is the desired* file structure:

1. hardware/interfaces/audio/11.0/*.hal
2. hardware/interfaces/audio/aidl/android/hardware/audio/*.aidl
3. hardware/interfaces/audio/aidl/Android.bp

* it is overly deep, but this keeps the aidl compiler consistent with
  javac.

So, to run hidl2aidl to generate these AIDL files, you should run:

    hidl2aidl -o hardware/interfaces/audio/aidl \
        android.hardware.audio@11.0

Bug: 162943848
Test: the following:
    source system/tools/hidl/scripts/hal-queries.sh
    for i in $(aosp-interfaces | cut -d: -f1 | sort -u); do hidl2aidl -f -o foo/$i/ $i; done
Change-Id: I754d39660a023c3042a0787c96ecb046eda57197