All HIDL tests device -> general-tests.
None of these need to be built per-device, only the architecture
matters.

This didn't work w/ general-tests before because tradefed was running
32-bit tests from test_suites_x86_64 on 32-bit devices, which caused
this test to fail, since it still installed 64-bit tests in that case,
and this test tried to run it. Although a separate bug is filed to
investigate that, switching the method of abi detection here.

Bug: 168807095
Test: runs tests on 32 and 64-bit devices
Change-Id: If7e5a952d9a94871b96bf8536870297e6272a956