hidl2aidl: prevent duplicate types before processing
Changing the vector to a set to drop any duplicate types.
These duplicate types were coming from an unnecessary call to
getSubTypes on interfaces. It is already being handled by calling it on
the typeChain for that interface that includes itself. This has been
moved to avoid the interfaces.
The second spot is coming from the traversal of the typeChain of an
interface that is extending a different interface. We get the subtypes
twice from that base interface. This behavior will not be changed.

Test: hidl2aidl -o out -rhidl2aidl.test:system/tools/hidl/hidl2aidl/test hidl2aidl@1.2
Change-Id: I5bc7b2f011158030d83c40517db6093aca254942