hidl2aidl: avoid platform_apis
When this build file was originally written, using platform_apis was the
only way to use @VintfStability interfaces. However, now, we can build
against module_current, which is the smallest stub API surface that
includes all AIDL dependencies. This also has the nice property that
modules depending on this stub can be included in the build of
frameworks.jar (whereas modules using platform_apis can't, because they
depend on frameworks.jar).

Bug: N/A
Test: run_build_file_test.sh
Change-Id: I5488654ebcc1f086757dc277e3d128aa3bd32986