Disallow new HIDL interfaces.
This saves a list of all HIDL interfaces and errors if new interfaces
are added. The Android T policy is that new AOSP interfaces must be
added in AIDL (ideally with full conversions, but optionally as
extension interfaces). This error is meant to avoid anyone accidentally
wasting time creating a new HIDL interface that cannot be accepted.

Now that is past the deadline for new Android S interfaces, adding this
in. Note - this contains interfaces that aren't on AOSP, but they are
visible in the android S preview. They are added here for consistency
only (this would allow people to easily cherry-pick S interfaces to AOSP
if desired).

Bug: 191781736
Fixes: 182941786
Test: m nothing
Change-Id: Ieadef0dea7d30ac5085f8a087b540755f2717a69
(cherry picked from commit 09b19ff2fe688c1ce995da967a82095bb87e3e3c)
Merged-In: Ieadef0dea7d30ac5085f8a087b540755f2717a69